Remove annotation on derived Schema
